Restructure components files
