+++ src/Components/App/App.css
@@ -1,13 +1,11 @@
 .App {
   text-align: center;
   font-family: 'Joan', serif;
-
 }
 
 
 h1 {
-  font-family: 'Joan', serif;
-  color:rgb(56, 16, 51); 
+  color: rgb(56, 16, 51);
   font-size: 10em;
   text-align: center;
 }
