+++ src/Components/Header/Header.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Navbar from '../Navbar/Navbar';
+import './Header.css';
+
+
+const Header = () => {
+    return (
+        <header className="header">
+          <h1>GÜT  BOOKS</h1>
+          < Navbar />
+        </header>
+    )
+}
+
+
+export default Header;
