+++ src/Components/Navbar/Navbar.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import '../styles/Navbar.css';
+import './Navbar.css';
 
 
 
 const Navbar = () => {
     return (
         <nav className="nav-bar">
-            <NavLink className="nav-link" to="/">
+            <NavLink className="nav-link" to='/'>
                 <p className='all-books'>All Books</p>
             </NavLink>
 
-            <NavLink className="nav-link" to="/saved">
+            <NavLink className="nav-link" to='/saved'>
                 <p className='my-shelf'>My Shelf</p>
             </NavLink>
         </nav>
