+++ src/Components/SaveShelf/SaveShelf.scss
@@ -4,7 +4,7 @@ $emerald: #1b7d54;
 $gold: #ecec71;
 
 
-.saved-wrapper {
+.save-shelf {
     width: 75%;
     height: 100%;
     margin: 0 auto;
