+++ src/index.css
@@ -8,9 +8,9 @@ body {
   margin: none;
 }
 
-header {
+/* header {
   background: url(books.jpeg) center fixed;
   background-size: cover;
   width: 100vw;
   padding-top: 1%;
-}
+} */
